requirements.txt

# Core
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2

# Deep learning
torch==2.3.0
torchvision==0.18.0

# Dimensionality reduction & viz
umap-learn==0.5.6
matplotlib==3.8.4
seaborn==0.13.2

# Utilities
tqdm==4.66.4
requests==2.32.2